pyproject.toml
[tool.poetry]
name = "crewai-sheets-ui"
version = "0.1.0"
description = ""
authors = ["Yuri"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.2.1"
python-dotenv = "<=1.0.0"
pydantic = "^2.6.4"
langchain = "^0.1.14"
crewai = {extras = ["tools"], version = "^0.22.5"}
duckduckgo-search = "^5.2.2"
open-interpreter = "^0.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
